Forgot to commit the database file
Database changes: Deleted no_show_flag in ea_user table, add that in
ea_appointment table. Created new column in ea_user as num_noshow.
created a box to display number of no show. The box is connected to the
database. Next step, to write a function count the number of no shows.